# internal/route/subscription.go
package route

import (
	"fmt"
	"net/http"
	"strconv"

	"github.com/ashwingopalsamy/uuidcheck"
	"github.com/gin-gonic/gin"
	"github.com/rs/zerolog/log"

	"subscription/internal/entities"
	utils "subscription/pkg"
)

type DPOSubscription struct {
	ServiceName string `json:"service_name"`
	Price       int    `json:"price"`
	UserUUID    string `json:"user_id"`
	StartDate   string `json:"start_date"`
	EndDate     string `json:"end_date"`
}

// HandleAddSubscription
// @Tags         Subscription
// @Summary      Добавление записи о подписке
// @Accept       json
// @Produce      json
// @Param        data body DPOSubscription true "Данные записи подписки"
// @Success      201 {object} entities.Subscription "Успешный ответ"
// @Failure      400 {object} entities.ErrorResponse "Некорректный запрос"
// @Failure      500 {object} entities.ErrorResponse "Ошибка на стороне сервера"
// @Router       /subscription [post]
func (h *Handler) HandleAddSubscription(ctx *gin.Context) {
	var dpo DPOSubscription
	err := ctx.BindJSON(&dpo)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: invalid request body"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: invalid request body")
		return
	}
	if dpo.ServiceName == "" {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: empty service name"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: empty service name")
		return
	}
	if dpo.Price == 0 {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: empty price"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: empty price")
		return
	}
	if dpo.Price < 0 {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: price is less than 0"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: price is less than 0")
		return
	}
	if dpo.UserUUID == "" {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: empty user uuid"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: empty user uuid")
		return
	}
	if !uuidcheck.IsValidUUID(dpo.UserUUID) {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: not valid user uuid"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: not valid user uuid")
		return
	}
	if dpo.StartDate == "" {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: empty start date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: empty start date")
		return
	}
	if !utils.IsValidMonthYear(dpo.StartDate) {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: wrong type start date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: wrong type start date")
		return
	}

	if dpo.EndDate != "" && !utils.IsValidMonthYear(dpo.EndDate) {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: wrong type end date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: wrong type end date")
		return
	}

	if dpo.EndDate != "" && !utils.IsValidMonthYearLength(dpo.StartDate, dpo.EndDate) {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: the end date is earlier than the start date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: the end date is earlier than the start date")
		return
	}

	var subscription entities.Subscription
	subscription.ServiceName = dpo.ServiceName
	subscription.Price = dpo.Price
	subscription.UserUUID = dpo.UserUUID
	subscription.StartDate, err = utils.MonthYearToDate(dpo.StartDate)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: wrong type start date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: wrong type start date")
		return
	}
	if dpo.EndDate != "" {
		subscription.EndDate, err = utils.MonthYearToDate(dpo.EndDate)
		if err != nil {
			ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: wrong type end date"})
			log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
				Msg("error getting request data: wrong type end date")
			return
		}
	}

	log.Debug().Msg("call h.services.Subscription.Add")
	outSubscription, err := h.services.Subscription.Add(&subscription)
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("error adding subscription: %s", err.Error())})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusInternalServerError).
			Msgf("error adding subscription: %s", err.Error())
		return
	}

	ctx.JSON(http.StatusCreated, outSubscription)
	log.Info().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusCreated).
		Msg("successful adding subscription")
}

// HandleGetAllSubscription
// @Tags         Subscription
// @Summary      Вывести все записи о подписках
// @Accept       json
// @Produce      json
// @Success      200 {object} []entities.Subscription "Успешный ответ"
// @Success      204 "Успешный ответ. Пустая база данных"
// @Failure      400 {object} entities.ErrorResponse "Некорректный запрос"
// @Failure      500 {object} entities.ErrorResponse "Ошибка на стороне сервера"
// @Router       /subscription [get]
func (h *Handler) HandleGetAllSubscription(ctx *gin.Context) {
	log.Debug().Msg("call h.services.Subscription.GetAll")
	outSubscriptionSlice, err := h.services.Subscription.GetAll()
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("error getting all subscriptions: %s", err.Error())})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusInternalServerError).
			Msgf("error getting all subscriptions: %s", err.Error())
		return
	}

	if len(*outSubscriptionSlice) == 0 {
		ctx.JSON(http.StatusNoContent, "")
		log.Info().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusOK).
			Msg("successful getting all subscriptions, but finding 0 rows")
		return
	}

	ctx.JSON(http.StatusOK, outSubscriptionSlice)
	log.Info().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusOK).
		Msg("successful getting all subscriptions")
}

// HandleGetSubscriptionById
// @Tags         Subscription
// @Summary      Вывести запись о подписке по id
// @Accept       json
// @Produce      json
// @Param        id  path      int true "ID записи о подписки" example(1)
// @Success      200 {object} entities.Subscription "Успешный ответ"
// @Failure      400 {object} entities.ErrorResponse "Некорректный запрос"
// @Failure      404 {object} entities.ErrorResponse "Запись не найдена"
// @Failure      500 {object} entities.ErrorResponse "Ошибка на стороне сервера"
// @Router       /subscription/{id} [get]
func (h *Handler) HandleGetSubscriptionById(ctx *gin.Context) {
	idStr := ctx.Param("id")
	if idStr == "" {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error parsing user data: empty id"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error parsing user data: empty id")
		return
	}
	id, err := strconv.Atoi(idStr)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error parsing user data: count is not a number"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error parsing user data: count is not a number")
		return
	}
	if id < 1 {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error parsing user data: invalid id"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error parsing user data: invalid id")
		return
	}

	log.Debug().Msg("call h.services.Subscription.GetById")
	outSubscription, err := h.services.Subscription.GetById(id)
	if err != nil {
		if err.Error() == "Subscription.GetById: sql: no rows in result set" {
			ctx.JSON(http.StatusNotFound, gin.H{"error": "error getting subscription by id: id not founded"})
			log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusNotFound).
				Msg("error getting subscription by id: id not founded")
			return
		}
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("error getting subscription by id: %s", err.Error())})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusInternalServerError).
			Msgf("error getting subscription by id: %s", err.Error())
		return
	}

	ctx.JSON(http.StatusOK, outSubscription)
	log.Info().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusOK).
		Msg("successful getting subscription by id")
}

// HandleGetSumSubscriptionByFilters
// @Tags         Subscription
// @Summary      Сумма подписок по фильтрам
// @Accept       json
// @Produce      json
// @Param        service_name query string false "Название сервиса"
// @Param        user_id query string false "UUID пользователя"
// @Param        start_date query string false "Дата начала (формат: MM-YYYY)"
// @Param        end_date query string false "Дата окончания (формат: MM-YYYY)"
// @Success      200 {object} entities.SubscriptionAmount "Успешный ответ"
// @Failure      400 {object} entities.ErrorResponse "Некорректный запрос"
// @Failure      500 {object} entities.ErrorResponse "Ошибка на стороне сервера"
// @Router       /subscription/filters [get]
func (h *Handler) HandleGetSumSubscriptionByFilters(ctx *gin.Context) {
	var filters entities.SubscriptionsFilters
	filters.ServiceName = ctx.Query("service_name")
	filters.UserUUID = ctx.Query("user_id")
	StartDateStr := ctx.Query("start_date")
	StartDate, err := utils.MonthYearToDate(StartDateStr)
	if StartDateStr != "" && err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: wrong type start date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: wrong type start date")
		return
	}
	filters.StartDate = StartDate

	EndDateStr := ctx.Query("end_date")
	EndDate, err := utils.MonthYearToDate(EndDateStr)
	if EndDateStr != "" && err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: wrong type end date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: wrong type end date")
		return
	}
	filters.EndDate = EndDate

	log.Debug().Msg("call h.services.Subscription.GetSumByFilters")
	var amount entities.SubscriptionAmount
	amount.Amount, err = h.services.Subscription.GetSumByFilters(&filters)
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("error getting sum subscriptions by filters: %s", err.Error())})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusInternalServerError).
			Msgf("error getting all subscriptions: %s", err.Error())
		return
	}

	ctx.JSON(http.StatusOK, amount)
	log.Info().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusOK).
		Msg("successful getting sum subscriptions by filters")
}

// HandleUpdateSubscription
// @Tags         Subscription
// @Summary      Изменить запись о подписке
// @Accept       json
// @Produce      json
// @Param        id  path      int true "ID записи о подписки" example(1)
// @Param        data body DPOSubscription true "Данные редактируемого записи"
// @Success      200 {object} entities.Subscription "Успешный ответ"
// @Failure      400 {object} entities.ErrorResponse "Некорректный запрос"
// @Failure      404 {object} entities.ErrorResponse "Запись не найдена"
// @Failure      500 {object} entities.ErrorResponse "Ошибка на стороне сервера"
// @Router       /subscription/{id} [put]
func (h *Handler) HandleUpdateSubscription(ctx *gin.Context) {
	idStr := ctx.Param("id")
	if idStr == "" {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error parsing user data: empty id"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error parsing user data: empty id")
		return
	}
	id, err := strconv.Atoi(idStr)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error parsing user data: count is not a number"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error parsing user data: count is not a number")
		return
	}
	if id < 1 {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error parsing user data: invalid id"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error parsing user data: invalid id")
		return
	}

	log.Debug().Msg("call h.services.Subscription.GetById")
	_, err = h.services.Subscription.GetById(id)
	if err != nil {
		if err.Error() == "Subscription.GetById: sql: no rows in result set" {
			ctx.JSON(http.StatusNotFound, gin.H{"error": "error getting subscription by id: id not founded"})
			log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusNotFound).
				Msg("error getting subscription by id: id not founded")
			return
		}
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("error getting subscription by id: %s", err.Error())})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusInternalServerError).
			Msgf("error getting subscription by id: %s", err.Error())
		return
	}

	var dpo DPOSubscription
	err = ctx.BindJSON(&dpo)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: invalid request body"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: invalid request body")
		return
	}
	if dpo.ServiceName == "" {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: empty service name"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: empty service name")
		return
	}
	if dpo.Price == 0 {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: empty price"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: empty price")
		return
	}
	if dpo.Price < 0 {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: price is less than 0"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: price is less than 0")
		return
	}
	if dpo.UserUUID == "" {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: empty user uuid"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: empty user uuid")
		return
	}
	if !uuidcheck.IsValidUUID(dpo.UserUUID) {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: not valid user uuid"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: not valid user uuid")
		return
	}
	if dpo.StartDate == "" {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: empty start date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: empty start date")
		return
	}
	if !utils.IsValidMonthYear(dpo.StartDate) {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: wrong type start date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: wrong type start date")
		return
	}

	if dpo.EndDate != "" && !utils.IsValidMonthYear(dpo.EndDate) {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: wrong type end date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: wrong type end date")
		return
	}

	if dpo.EndDate != "" && !utils.IsValidMonthYearLength(dpo.StartDate, dpo.EndDate) {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: the end date is earlier than the start date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: the end date is earlier than the start date")
		return
	}

	var subscription entities.Subscription
	subscription.Id = id
	subscription.ServiceName = dpo.ServiceName
	subscription.Price = dpo.Price
	subscription.UserUUID = dpo.UserUUID
	subscription.StartDate, err = utils.MonthYearToDate(dpo.StartDate)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: wrong type start date"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error getting request data: wrong type start date")
		return
	}
	if dpo.EndDate != "" {
		subscription.EndDate, err = utils.MonthYearToDate(dpo.EndDate)
		if err != nil {
			ctx.JSON(http.StatusBadRequest, gin.H{"error": "error getting request data: wrong type end date"})
			log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
				Msg("error getting request data: wrong type end date")
			return
		}
	}

	log.Debug().Msg("call h.services.Subscription.Update")
	outSubscription, err := h.services.Subscription.Update(&subscription)
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("error updateting subscription: %s", err.Error())})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusInternalServerError).
			Msgf("error updateting subscription: %s", err.Error())
		return
	}

	ctx.JSON(http.StatusOK, outSubscription)
	log.Info().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusOK).
		Msg("successful updateting subscription")
}

// HandleDeleteSubscription
// @Tags         Subscription
// @Summary      Удалить запись о подписке
// @Accept       json
// @Produce      json
// @Param        id  path      int true "ID записи о подписки" example(1)
// @Success      200 {object} entities.Message "Успешный ответ"
// @Failure      400 {object} entities.ErrorResponse "Некорректный запрос"
// @Failure      404 {object} entities.ErrorResponse "Запись не найдена"
// @Failure      500 {object} entities.ErrorResponse "Ошибка на стороне сервера"
// @Router       /subscription/{id} [delete]
func (h *Handler) HandleDeleteSubscription(ctx *gin.Context) {
	idStr := ctx.Param("id")
	if idStr == "" {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error parsing user data: empty id"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error parsing user data: empty id")
		return
	}
	id, err := strconv.Atoi(idStr)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error parsing user data: count is not a number"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error parsing user data: count is not a number")
		return
	}
	if id < 1 {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error parsing user data: invalid id"})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusBadRequest).
			Msg("error parsing user data: invalid id")
		return
	}

	log.Debug().Msg("call h.services.Subscription.GetById")
	_, err = h.services.Subscription.GetById(id)
	if err != nil {
		if err.Error() == "Subscription.GetById: sql: no rows in result set" {
			ctx.JSON(http.StatusNotFound, gin.H{"error": "error getting subscription by id: id not founded"})
			log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusNotFound).
				Msg("error getting subscription by id: id not founded")
			return
		}
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("error getting subscription by id: %s", err.Error())})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusInternalServerError).
			Msgf("error getting subscription by id: %s", err.Error())
		return
	}

	log.Debug().Msg("call h.services.Subscription.GetById")
	err = h.services.Subscription.Delete(id)

	if err != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("error deleting subscription: %s", err.Error())})
		log.Error().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusInternalServerError).
			Msgf("error deleting subscription: %s", err.Error())
		return
	}

	ctx.JSON(http.StatusOK, gin.H{"message": "ok"})
	log.Info().Str("method", ctx.Request.Method).Str("url", ctx.Request.URL.String()).Int("status", http.StatusOK).
		Msg("successful deleting subscription")
}